# Fix `no-descending-specificity` performance