--- a/.changeset/thirty-swans-complain.md
+++ b/.changeset/thirty-swans-complain.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `no-descending-specificity` performance