--- a/lib/rules/no-descending-specificity/__tests__/index.js
+++ b/lib/rules/no-descending-specificity/__tests__/index.js
@@ -190,6 +190,58 @@ testRule({
 			endLine: 1,
 			endColumn: 29,
 		},
+		{
+			code: '.a .b .c {} .b .c {} .c {}',
+			warnings: [
+				{
+					message: messages.rejected('.b .c', '.a .b .c'),
+					line: 1,
+					column: 13,
+					endLine: 1,
+					endColumn: 18,
+				},
+				{
+					message: messages.rejected('.c', '.a .b .c'),
+					line: 1,
+					column: 22,
+					endLine: 1,
+					endColumn: 24,
+				},
+			],
+		},
+		{
+			code: '.a .b .c {} .b .c {} .c {} .d .b .c {} .b .c {} .c {}',
+			warnings: [
+				{
+					message: messages.rejected('.b .c', '.a .b .c'),
+					line: 1,
+					column: 13,
+					endLine: 1,
+					endColumn: 18,
+				},
+				{
+					message: messages.rejected('.c', '.a .b .c'),
+					line: 1,
+					column: 22,
+					endLine: 1,
+					endColumn: 24,
+				},
+				{
+					message: messages.rejected('.b .c', '.a .b .c'),
+					line: 1,
+					column: 40,
+					endLine: 1,
+					endColumn: 45,
+				},
+				{
+					message: messages.rejected('.c', '.a .b .c'),
+					line: 1,
+					column: 49,
+					endLine: 1,
+					endColumn: 51,
+				},
+			],
+		},
 	],
 });