--- a/lib/rules/no-descending-specificity/index.js
+++ b/lib/rules/no-descending-specificity/index.js
@@ -13,7 +13,6 @@ const parseSelector = require('../../utils/parseSelector');
 const report = require('../../utils/report');
 const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
-const { assert } = require('../../utils/validateTypes');
 
 const ruleName = 'no-descending-specificity';
 
@@ -63,8 +62,10 @@ const rule = (primary, secondaryOptions) => {
 				return;
 			}
 
+			const selectors = ruleNode.selectors;
+
 			// Ignores selectors within list of selectors
-			if (ignoreSelectorsWithinList && ruleNode.selectors.length > 1) {
+			if (ignoreSelectorsWithinList && selectors.length > 1) {
 				return;
 			}
 
@@ -74,43 +75,41 @@ const rule = (primary, secondaryOptions) => {
 				findAtRuleContext(ruleNode),
 			);
 
-			for (const selector of ruleNode.selectors) {
-				const trimSelector = selector.trim();
-
+			for (const selector of selectors) {
 				// Ignore `.selector, { }`
-				if (trimSelector === '') {
+				if (selector.trim() === '') {
 					continue;
 				}
 
 				// Resolve any nested selectors before checking
 				for (const resolvedSelector of resolvedNestedSelector(selector, ruleNode)) {
-					parseSelector(resolvedSelector, result, ruleNode, (s) => {
-						if (!isStandardSyntaxSelector(resolvedSelector)) {
-							return;
-						}
+					if (!isStandardSyntaxSelector(resolvedSelector)) {
+						continue;
+					}
 
-						checkSelector(s, ruleNode, comparisonContext);
+					parseSelector(resolvedSelector, result, ruleNode, (s) => {
+						checkSelector(resolvedSelector, s, ruleNode, comparisonContext);
 					});
 				}
 			}
 		});
 
 		/**
+		 * @param {string} selector
 		 * @param {import('postcss-selector-parser').Root} selectorNode
 		 * @param {import('postcss').Rule} ruleNode
 		 * @param {Map<string, Entry[]>} comparisonContext
 		 */
-		function checkSelector(selectorNode, ruleNode, comparisonContext) {
-			const selector = selectorNode.toString();
+		function checkSelector(selector, selectorNode, ruleNode, comparisonContext) {
 			const referenceSelector = lastCompoundSelectorWithoutPseudoClasses(selectorNode);
 
-			if (referenceSelector === undefined) return;
+			if (!referenceSelector) return;
 
 			const selectorSpecificity = calculate(selectorNode);
 			const entry = { selector, specificity: selectorSpecificity };
 			const priorComparableSelectors = comparisonContext.get(referenceSelector);
 
-			if (priorComparableSelectors === undefined) {
+			if (!priorComparableSelectors) {
 				comparisonContext.set(referenceSelector, [entry]);
 
 				return;
@@ -126,6 +125,8 @@ const rule = (primary, secondaryOptions) => {
 						messageArgs: [selector, priorEntry.selector],
 						word: selector,
 					});
+
+					break;
 				}
 			}
 
@@ -141,11 +142,12 @@ const rule = (primary, secondaryOptions) => {
 function lastCompoundSelectorWithoutPseudoClasses(selectorNode) {
 	const firstChild = selectorNode.nodes[0];
 
-	assert(firstChild);
+	if (!firstChild) return undefined;
+
 	const nodesByCombinator = firstChild.split((node) => node.type === 'combinator');
 	const nodesAfterLastCombinator = nodesByCombinator[nodesByCombinator.length - 1];
 
-	assert(nodesAfterLastCombinator);
+	if (!nodesAfterLastCombinator) return undefined;
 
 	const nodesWithoutPseudoClasses = nodesAfterLastCombinator.filter((node) => {
 		return (